fix: rename file and router naming fix
