+++ src/frontend/app/src/components/navbar/NavBar.vue
@@ -55,7 +55,7 @@ export default defineComponent({
         { path: "/2fa", message: "Two Factor" },
         { path: "/login", message: "Login" },
         { path: "/pong", message: "Pong" },
-        { path: "/friend-view", message: "friend-view" },
+        { path: "/friend-list", message: "friend-list" },
       ],
     };
   },
