+++ src/frontend/app/src/views/FriendListView.vue
